[hollow] Fix: rework structure to provider directory version of the Hollow root. Do not forget any delta file less than 1 minute old. Use nio watch service to check local annoucements.

+++ parent/common/net.bluemind.common.hollow/src/net/bluemind/common/hollow/BmFilesystemBlobStorageCleaner.java
@@ -20,6 +20,7 @@ package net.bluemind.common.hollow;
 import java.io.File;
 import java.util.Arrays;
 import java.util.Collections;
+import java.util.concurrent.TimeUnit;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -70,8 +71,14 @@ public class BmFilesystemBlobStorageCleaner extends HollowProducer.BlobStorageCl
 
 		sortByLastModified(files);
 		int del = 0;
+		long now = System.currentTimeMillis();
 		for (int i = kept; i < files.length; i++) {
 			File file = files[i];
+
+			if (now - file.lastModified() < TimeUnit.MINUTES.toMillis(1)) {
+				// when doing fast writes, leave room to the consumers to catch up
+				continue;
+			}
 			boolean deleted = file.delete(); // NOSONAR
 			if (!deleted) {
 				logger.warn("Could not delete delta {}", file.getPath());
